remove weird cast
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/branches/3.0.x@1175847 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/ClusterHandler.cc
@@ -1863,7 +1863,7 @@ ClusterHandler::add_small_controlmsg_descriptors()
     c->freeall();
     p = (char *) DOUBLE_ALIGN(p);
 #ifdef PURIFY
-    if ((unsigned int) endp < (unsigned int) p)
+    if (endp < p)
       memset(endp, 0, (p - endp));
 #endif
   }
